Move LayeredModel to a new file.
